requirements.txt
tensorflow-macos==2.13.0
tensorflow-metal==1.0.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.2.1
nltk==3.8.1
flask==3.0.2
flask-cors==4.0.0
joblib==1.3.2
matplotlib==3.8.3
seaborn==0.13.2
Pillow==10.2.0
opencv-python==4.9.0.80
moviepy==1.0.3
imageio==2.9.0
imageio-ffmpeg==0.4.7
decorator==4.4.2
setuptools==67.8.0
fer==22.5.1
gunicorn==21.2.0